{{cookiecutter.project_name}}/requirements.txt
pytest
python-dotenv
